requirements.txt
openai
requests
selenium
python-dotenv
beautifulsoup4
pandas
Pillow
flask
flask-cors
openpyxl